refactor: move payment to domain module
