+++ src/domain/mod.rs
@@ -1,1 +1,2 @@
 pub mod income;
+pub mod payment;
